+++ src/app/(pages)/(auth)/signin/page.tsx
@@ -1,0 +1,180 @@
+"use client";
+
+import type React from "react";
+
+import { useState } from "react";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Alert, AlertDescription } from "@/components/ui/alert";
+import { Eye, EyeOff, Shield, User, Lock, Building } from "lucide-react";
+import { Button } from "@/components/ui/button";
+
+interface LoginPageProps {
+  onLogin: (username: string) => void;
+}
+
+function SigninPage({ onLogin }: LoginPageProps) {
+  const [formData, setFormData] = useState({
+    username: "",
+    password: "",
+  });
+  const [showPassword, setShowPassword] = useState(false);
+  const [error, setError] = useState("");
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setError("");
+    setIsLoading(true);
+
+    // 로그인 시뮬레이션 (1초 대기)
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+
+    // 계정 검증
+    if (formData.username === "admin" && formData.password === "admin") {
+      onLogin(formData.username);
+    } else {
+      setError("아이디 또는 비밀번호가 올바르지 않습니다.");
+    }
+
+    setIsLoading(false);
+  };
+
+  const handleInputChange = (field: string, value: string) => {
+    setFormData((prev) => ({ ...prev, [field]: value }));
+    if (error) setError(""); // 입력 시 에러 메시지 제거
+  };
+
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-4">
+      <div className="w-full max-w-md">
+        {/* 로고 및 헤더 */}
+        <div className="text-center mb-8">
+          <div className="inline-flex items-center justify-center w-16 h-16 bg-blue-600 rounded-full mb-4">
+            <Shield className="h-8 w-8 text-white" />
+          </div>
+          <h1 className="text-3xl font-bold text-gray-900 mb-2">
+            보험 설계사 CRM
+          </h1>
+          <p className="text-gray-600">고객 관리 시스템에 로그인하세요</p>
+        </div>
+
+        {/* 로그인 폼 */}
+        <Card className="shadow-lg">
+          <CardHeader className="space-y-1">
+            <CardTitle className="text-2xl text-center">로그인</CardTitle>
+            <CardDescription className="text-center">
+              계정 정보를 입력하여 시스템에 접속하세요
+            </CardDescription>
+          </CardHeader>
+          <CardContent>
+            <form onSubmit={handleSubmit} className="space-y-4">
+              {/* 아이디 입력 */}
+              <div className="space-y-2">
+                <Label htmlFor="username">아이디</Label>
+                <div className="relative">
+                  <User className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-4 w-4" />
+                  <Input
+                    id="username"
+                    type="text"
+                    placeholder="아이디를 입력하세요"
+                    value={formData.username}
+                    onChange={(e) =>
+                      handleInputChange("username", e.target.value)
+                    }
+                    className="pl-10"
+                    required
+                  />
+                </div>
+              </div>
+
+              {/* 비밀번호 입력 */}
+              <div className="space-y-2">
+                <Label htmlFor="password">비밀번호</Label>
+                <div className="relative">
+                  <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 h-4 w-4" />
+                  <Input
+                    id="password"
+                    type={showPassword ? "text" : "password"}
+                    placeholder="비밀번호를 입력하세요"
+                    value={formData.password}
+                    onChange={(e) =>
+                      handleInputChange("password", e.target.value)
+                    }
+                    className="pl-10 pr-10"
+                    required
+                  />
+                  <button
+                    type="button"
+                    onClick={() => setShowPassword(!showPassword)}
+                    className="absolute right-3 top-1/2 transform -translate-y-1/2 text-gray-400 hover:text-gray-600"
+                  >
+                    {showPassword ? (
+                      <EyeOff className="h-4 w-4" />
+                    ) : (
+                      <Eye className="h-4 w-4" />
+                    )}
+                  </button>
+                </div>
+              </div>
+
+              {/* 에러 메시지 */}
+              {error && (
+                <Alert variant="destructive">
+                  <AlertDescription>{error}</AlertDescription>
+                </Alert>
+              )}
+
+              {/* 로그인 버튼 */}
+              <Button type="submit" className="w-full" disabled={isLoading}>
+                {isLoading ? (
+                  <>
+                    <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
+                    로그인 중...
+                  </>
+                ) : (
+                  "로그인"
+                )}
+              </Button>
+            </form>
+
+            {/* 데모 계정 안내 */}
+            <div className="mt-6 p-4 bg-blue-50 rounded-lg">
+              <div className="flex items-center gap-2 mb-2">
+                <Building className="h-4 w-4 text-blue-600" />
+                <span className="text-sm font-medium text-blue-800">
+                  데모 계정 정보
+                </span>
+              </div>
+              <div className="text-sm text-blue-700 space-y-1">
+                <p>
+                  • 아이디:{" "}
+                  <code className="bg-blue-100 px-1 rounded">admin</code>
+                </p>
+                <p>
+                  • 비밀번호:{" "}
+                  <code className="bg-blue-100 px-1 rounded">admin</code>
+                </p>
+              </div>
+            </div>
+          </CardContent>
+        </Card>
+
+        {/* 푸터 */}
+        <div className="text-center mt-8 text-sm text-gray-500">
+          <p>© 2024 보험 설계사 CRM. All rights reserved.</p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default SigninPage;
