+++ src/app/data.ts
@@ -1,0 +1,77 @@
+// 샘플 데이터
+export const sampleCustomers = [
+  {
+    id: 1,
+    name: "김철수",
+    phone: "010-1234-5678",
+    birthDate: "1985-03-15",
+    joinDate: "2023-01-15",
+    renewalDate: "2024-03-15",
+    product: "종신보험",
+    paymentMethod: "자동이체",
+    premium: 150000,
+    age: 39,
+    memo: "정기 상담 필요",
+  },
+  {
+    id: 2,
+    name: "이영희",
+    phone: "010-9876-5432",
+    birthDate: "1990-07-22",
+    joinDate: "2023-06-10",
+    renewalDate: "2024-07-22",
+    product: "건강보험",
+    paymentMethod: "카드결제",
+    premium: 80000,
+    age: 34,
+    memo: "건강검진 결과 확인 필요",
+  },
+  {
+    id: 3,
+    name: "박민수",
+    phone: "010-5555-1234",
+    birthDate: "1978-11-08",
+    joinDate: "2022-11-08",
+    renewalDate: "2024-11-08",
+    product: "연금보험",
+    paymentMethod: "자동이체",
+    premium: 200000,
+    age: 46,
+    memo: "연금 수령 상담 예정",
+  },
+];
+
+export const todayTasks = [
+  {
+    id: 1,
+    type: "renewal",
+    title: "김철수님 상령일 D-30",
+    description: "종신보험 갱신 상담 필요",
+    priority: "high",
+    dueDate: "2024-02-15",
+  },
+  {
+    id: 2,
+    type: "birthday",
+    title: "이영희님 생일",
+    description: "생일 축하 연락",
+    priority: "medium",
+    dueDate: "2024-01-15",
+  },
+  {
+    id: 3,
+    type: "care",
+    title: "박민수님 가입 90일 케어",
+    description: "가입 후 첫 상담",
+    priority: "medium",
+    dueDate: "2024-01-15",
+  },
+  {
+    id: 4,
+    type: "followup",
+    title: "최지은님 미응답",
+    description: "3일 전 연락 후 미응답",
+    priority: "low",
+    dueDate: "2024-01-12",
+  },
+];
